Optimize CRUD dialog
