+++ lib/ui/expense/items/items_list.dart
@@ -8,7 +8,7 @@ import 'package:statera/data/models/item.dart';
 import 'package:statera/ui/expense/expense_action_handlers.dart';
 import 'package:statera/ui/expense/expense_builder.dart';
 import 'package:statera/ui/expense/items/item_list_item.dart';
-import 'package:statera/ui/widgets/dialogs/crud_dialog.dart';
+import 'package:statera/ui/widgets/dialogs/crud_dialog/crud_dialog.dart';
 import 'package:statera/ui/widgets/list_empty.dart';
 import 'package:statera/ui/widgets/optionally_dismissible.dart';
 import 'package:statera/utils/utils.dart';
