+++ lib/ui/widgets/dialogs/dialogs.dart
@@ -1,3 +1,3 @@
-export 'crud_dialog.dart';
+export 'crud_dialog/crud_dialog.dart';
 export 'ok_cancel_dialog.dart';
 export 'payment_dialog.dart';
